refactor: extract optimistic update hooks
