+++ src/Components/Common/Card/subcomponents/CardBack/CardBack.tsx
@@ -1,6 +1,9 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useAtomInput, useCardSide, useCorrect } from '../../../../../hooks';
-import { updateCardsAPI } from '../../../../../api/cardClient';
+import {
+  useAtomInput,
+  useCardMutation,
+  useCardSide,
+  useCorrect,
+} from '@/hooks';
 import {
   AnswerContainer,
   CardBackContainer,
@@ -17,34 +20,8 @@ type CardBackProps = {
 };
 
 export function CardBack({ _id, answer, question, stackCount }: CardBackProps) {
-  const queryClient = useQueryClient();
-  const { mutate: updateCard } = useMutation({
-    mutationFn: updateCardsAPI,
-    onMutate: async (cardItem) => {
-      await queryClient.cancelQueries({ queryKey: ['cards'] });
-
-      const previousCards: Card[] = queryClient.getQueryData(['cards']) ?? [];
-
-      queryClient.setQueryData<Card[]>(['cards'], (oldCards) => {
-        if (oldCards) {
-          return [...oldCards].map((card) =>
-            card._id === cardItem.id
-              ? { _id: cardItem.id, ...cardItem.card }
-              : card
-          );
-        } else return [];
-      });
-      return { previousCards };
-    },
-    onError: (_err, _cardItem, context) => {
-      if (context) queryClient.setQueryData(['cards'], context.previousCards);
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ['cards'] });
-    },
-  });
-
   const { cardSide, toggleTo } = useCardSide();
+  const { updateCard } = useCardMutation();
 
   const { isCorrect } = useCorrect();
   const { inputVal, resetInputVal } = useAtomInput();
