+++ src/Components/Common/Card/subcomponents/EditCard/EditCard.tsx
@@ -1,6 +1,4 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useCardSide, useInput } from '../../../../../hooks';
-import { updateCardsAPI } from '../../../../../api/cardClient';
+import { useCardMutation, useCardSide, useInput } from '@/hooks';
 import { Button, Input } from '../../..';
 import { CardEditContainer } from './EditCard.style';
 
@@ -15,32 +13,7 @@ export function EditCard({
 }: EditCardProps) {
   const { cardSide, togglePrev } = useCardSide();
 
-  const queryClient = useQueryClient();
-  const { mutate: updateCard } = useMutation({
-    mutationFn: updateCardsAPI,
-    onMutate: async (cardItem) => {
-      await queryClient.cancelQueries({ queryKey: ['cards'] });
-
-      const previousCards: Card[] = queryClient.getQueryData(['cards']) ?? [];
-
-      queryClient.setQueryData<Card[]>(['cards'], (oldCards) => {
-        if (oldCards)
-          return [...oldCards].map((card) =>
-            card._id === cardItem.id
-              ? { _id: cardItem.id, ...cardItem.card }
-              : card
-          );
-        else return [];
-      });
-      return { previousCards };
-    },
-    onError: (_err, _cardItem, context) => {
-      if (context) queryClient.setQueryData(['cards'], context.previousCards);
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ['cards'] });
-    },
-  });
+  const { updateCard } = useCardMutation();
 
   const {
     inputVal: questionVal,
